Accept review suggestion from Cristian
Interesting part to me is that websocket notifications only arrive on the session/websocket that request was made on. Or is it per session and not WebSocket? For example can I (for whatever non-logical reason) open a Websocket and register for event notification on other WebSocket connections/sessions?

Accepting the changes but curious about how it works.

Co-authored-by: cristian-recoseanu <cristian.recoseanu@hotmail.com>